Estimate Standardized Regression Coefficients and Sampling Covariance Matrix Assuming Multivariate Normality

Estimate Standardized Regression Coefficients and Sampling Covariance Matrix Assuming Multivariate Normality

Usage

BetaN(object)

Arguments

object

Object of class lm.

Value

Returns an object of class betaSandwich which is a list with the following elements:

call

Function call.

type

Standard error type.

beta

Vector of standardized slopes.

vcov

Sampling covariance matrix of the standardized slopes.

n

Sample size.

p

Number of regressors.

df

\(n - p - 1\) degrees of freedom

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaN(object)
# Methods -------------------------------------------------------
print(std)
#> Call:
#> BetaN(object = object)
#> 
#> Standardized regression slopes with MVN standard errors:
#>            est     se      t     p   0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> NARTIC  0.4951 0.0759 6.5272 0.000  0.2268 0.2905 0.3421 0.6482 0.6998 0.7635
#> PCTGRT  0.3915 0.0770 5.0824 0.000  0.1190 0.1837 0.2360 0.5469 0.5993 0.6640
#> PCTSUPP 0.2632 0.0747 3.5224 0.001 -0.0011 0.0616 0.1124 0.4141 0.4649 0.5276
summary(std)
#> Call:
#> BetaN(object = object)
#> 
#> Standardized regression slopes with MVN standard errors:
#>            est     se      t     p   0.05%   0.5%   2.5%  97.5%  99.5% 99.95%
#> NARTIC  0.4951 0.0759 6.5272 0.000  0.2268 0.2905 0.3421 0.6482 0.6998 0.7635
#> PCTGRT  0.3915 0.0770 5.0824 0.000  0.1190 0.1837 0.2360 0.5469 0.5993 0.6640
#> PCTSUPP 0.2632 0.0747 3.5224 0.001 -0.0011 0.0616 0.1124 0.4141 0.4649 0.5276
coef(std)
#>    NARTIC    PCTGRT   PCTSUPP 
#> 0.4951451 0.3914887 0.2632477 
vcov(std)
#>               NARTIC       PCTGRT      PCTSUPP
#> NARTIC   0.005754524 -0.003360334 -0.002166127
#> PCTGRT  -0.003360334  0.005933462 -0.001769723
#> PCTSUPP -0.002166127 -0.001769723  0.005585256
confint(std, level = 0.95)
#>              2.5%     97.5%
#> NARTIC  0.3420563 0.6482339
#> PCTGRT  0.2360380 0.5469395
#> PCTSUPP 0.1124272 0.4140682