Estimate Standardized Regression Coefficients and Robust Sampling Covariance Matrix

Estimate Standardized Regression Coefficients and Robust Sampling Covariance Matrix

Usage

BetaHC(object, type = "hc3", g1 = 1, g2 = 1.5, k = 0.7)

Arguments

object

Object of class lm.

type

Character string. Correction type. Possible values are "hc0", "hc1", "hc2", "hc3", "hc4", "hc4m", and "hc5".

g1

Numeric. g1 value for type = "hc4m" or type = "hc5".

g2

Numeric. g2 value for type = "hc4m".

k

Numeric. Constant for type = "hc5"

Value

Returns an object of class betaSandwich which is a list with the following elements:

call

Function call.

type

Standard error type.

beta

Vector of standardized slopes.

vcov

Sampling covariance matrix of the standardized slopes.

n

Sample size.

p

Number of regressors.

df

\(n - p - 1\) degrees of freedom

References

Dudgeon, P. (2017). Some improvements in confidence intervals for standardized regression coefficients. Psychometrika, 82(4), 928–951. doi:10.1007/s11336-017-9563-z

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

object <- lm(rating ~ ., data = attitude)
std <- BetaHC(object)
# Methods -------------------------------------------------------
print(std)
#> Call:
#> BetaHC(object = object)
#> 
#> Standardized regression slopes with HC3 standard errors:
#>                 est     se       t         p    0.05%    0.5%     2.5%  97.5%
#> complaints  0.67073 0.1560  4.2983 0.0002677  0.08281  0.2327  0.34792 0.9935
#> privileges -0.07343 0.1594 -0.4608 0.6493035 -0.67385 -0.5208 -0.40310 0.2562
#> learning    0.30887 0.1954  1.5806 0.1276145 -0.42735 -0.2397 -0.09536 0.7131
#> raises      0.06981 0.1843  0.3787 0.7083821 -0.62473 -0.4477 -0.31153 0.4512
#> critical    0.03120 0.1772  0.1761 0.8617538 -0.63630 -0.4662 -0.33530 0.3977
#> advance    -0.18346 0.1802 -1.0183 0.3191381 -0.86228 -0.6893 -0.55618 0.1892
#>             99.5% 99.95%
#> complaints 1.1088 1.2586
#> privileges 0.3740 0.5270
#> learning   0.8574 1.0451
#> raises     0.5873 0.7644
#> critical   0.5286 0.6987
#> advance    0.3223 0.4954
summary(std)
#> Call:
#> BetaHC(object = object)
#> 
#> Standardized regression slopes with HC3 standard errors:
#>                    est        se          t            p       0.05%       0.5%
#> complaints  0.67072520 0.1560447  4.2982892 0.0002677147  0.08280705  0.2326554
#> privileges -0.07342743 0.1593643 -0.4607521 0.6493034784 -0.67385252 -0.5208164
#> learning    0.30887024 0.1954076  1.5806464 0.1276144566 -0.42735248 -0.2397043
#> raises      0.06981172 0.1843445  0.3787025 0.7083820908 -0.62472951 -0.4477051
#> critical    0.03119975 0.1771669  0.1761037 0.8617538415 -0.63629904 -0.4661672
#> advance    -0.18346445 0.1801712 -1.0182783 0.3191381252 -0.86228242 -0.6892656
#>                   2.5%     97.5%     99.5%    99.95%
#> complaints  0.34792217 0.9935282 1.1087950 1.2586434
#> privileges -0.40309753 0.2562427 0.3739616 0.5269977
#> learning   -0.09536107 0.7131016 0.8574448 1.0450930
#> raises     -0.31153391 0.4511573 0.5873286 0.7643529
#> critical   -0.33529792 0.3976974 0.5285667 0.6986985
#> advance    -0.55617704 0.1892481 0.3223367 0.4953535
coef(std)
#>  complaints  privileges    learning      raises    critical     advance 
#>  0.67072520 -0.07342743  0.30887024  0.06981172  0.03119975 -0.18346445 
vcov(std)
#>              complaints    privileges      learning       raises     critical
#> complaints  0.024349945 -0.0134050194 -0.0128301528 -0.012214863  0.002279143
#> privileges -0.013405019  0.0253969700 -0.0007948184  0.005897181 -0.004584845
#> learning   -0.012830153 -0.0007948184  0.0381841112 -0.013696467  0.004818864
#> raises     -0.012214863  0.0058971811 -0.0136964673  0.033982889 -0.010859760
#> critical    0.002279143 -0.0045848451  0.0048188637 -0.010859760  0.031388113
#> advance     0.010598636 -0.0100479388 -0.0083331127 -0.014248753  0.001774910
#>                 advance
#> complaints  0.010598636
#> privileges -0.010047939
#> learning   -0.008333113
#> raises     -0.014248753
#> critical    0.001774910
#> advance     0.032461673
confint(std, level = 0.95)
#>                   2.5%     97.5%
#> complaints  0.34792217 0.9935282
#> privileges -0.40309753 0.2562427
#> learning   -0.09536107 0.7131016
#> raises     -0.31153391 0.4511573
#> critical   -0.33529792 0.3976974
#> advance    -0.55617704 0.1892481