Estimate Standardized Regression Coefficients and Sampling Covariance Matrix Assuming Multivariate Normality

Estimate Standardized Regression Coefficients and Sampling Covariance Matrix Assuming Multivariate Normality

Usage

BetaN(object)

Arguments

object

Object of class lm.

Value

Returns an object of class betaSandwich which is a list with the following elements:

call

Function call.

type

Standard error type.

beta

Vector of standardized slopes.

vcov

Sampling covariance matrix of the standardized slopes.

n

Sample size.

p

Number of regressors.

df

\(n - p - 1\) degrees of freedom

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

object <- lm(rating ~ ., data = attitude)
std <- BetaN(object)
# Methods -------------------------------------------------------
print(std)
#> Call:
#> BetaN(object = object)
#> 
#> Standardized regression slopes with MVN standard errors:
#>                 est     se       t         p   0.05%     0.5%     2.5%   97.5%
#> complaints  0.67073 0.1433  4.6810 0.0001031  0.1309  0.26847  0.37431 0.96714
#> privileges -0.07343 0.1197 -0.6136 0.5455017 -0.5243 -0.40937 -0.32098 0.17412
#> learning    0.30887 0.1431  2.1580 0.0416048 -0.2304 -0.09293  0.01279 0.60495
#> raises      0.06981 0.1657  0.4213 0.6774348 -0.5545 -0.39536 -0.27296 0.41259
#> critical    0.03120 0.1047  0.2980 0.7683711 -0.3632 -0.26271 -0.18538 0.24778
#> advance    -0.18346 0.1338 -1.3717 0.1834027 -0.6874 -0.55896 -0.46015 0.09323
#>             99.5% 99.95%
#> complaints 1.0730 1.2106
#> privileges 0.2625 0.3774
#> learning   0.7107 0.8481
#> raises     0.5350 0.6941
#> critical   0.3251 0.4256
#> advance    0.1920 0.3205
summary(std)
#> Call:
#> BetaN(object = object)
#> 
#> Standardized regression slopes with MVN standard errors:
#>                    est        se          t            p      0.05%        0.5%
#> complaints  0.67072520 0.1432881  4.6809545 0.0001030949  0.1308690  0.26846733
#> privileges -0.07342743 0.1196668 -0.6135989 0.5455016623 -0.5242873 -0.40937234
#> learning    0.30887024 0.1431252  2.1580422 0.0416047589 -0.2303721 -0.09293027
#> raises      0.06981172 0.1656987  0.4213173 0.6774347660 -0.5544790 -0.39536008
#> critical    0.03119975 0.1046937  0.2980099 0.7683710623 -0.3632469 -0.26271051
#> advance    -0.18346445 0.1337535 -1.3716605 0.1834027121 -0.6873979 -0.55895555
#>                   2.5%      97.5%     99.5%    99.95%
#> complaints  0.37431113 0.96713928 1.0729831 1.2105814
#> privileges -0.32097709 0.17412223 0.2625175 0.3774325
#> learning    0.01279319 0.60494730 0.7106708 0.8481126
#> raises     -0.27296210 0.41258553 0.5349835 0.6941025
#> critical   -0.18537560 0.24777510 0.3251100 0.4256464
#> advance    -0.46015474 0.09322584 0.1920266 0.3204690
coef(std)
#>  complaints  privileges    learning      raises    critical     advance 
#>  0.67072520 -0.07342743  0.30887024  0.06981172  0.03119975 -0.18346445 
vcov(std)
#>               complaints    privileges     learning       raises      critical
#> complaints  0.0205314876 -0.0063811296 -0.009324286 -0.013811718 -0.0002422133
#> privileges -0.0063811296  0.0143201460 -0.002170471  0.001552377 -0.0002768442
#> learning   -0.0093242861 -0.0021704714  0.020484826 -0.004998152  0.0028372586
#> raises     -0.0138117179  0.0015523774 -0.004998152  0.027456049 -0.0048713593
#> critical   -0.0002422133 -0.0002768442  0.002837259 -0.004871359  0.0109607636
#> advance     0.0096650976 -0.0029432354 -0.006326814 -0.009305030 -0.0017608366
#>                 advance
#> complaints  0.009665098
#> privileges -0.002943235
#> learning   -0.006326814
#> raises     -0.009305030
#> critical   -0.001760837
#> advance     0.017890011
confint(std, level = 0.95)
#>                   2.5%      97.5%
#> complaints  0.37431113 0.96713928
#> privileges -0.32097709 0.17412223
#> learning    0.01279319 0.60494730
#> raises     -0.27296210 0.41258553
#> critical   -0.18537560 0.24777510
#> advance    -0.46015474 0.09322584